Bug 25641: Fix Koha/XSLT/Base.t failures on U20
There is a change in the version number.
On U20 we get
  Opening and ending tag mismatch: a line 0 and b
On others (D9, D10, U18) we get:
  Opening and ending tag mismatch: a line 1 and b

I did not find where it did come from, so assuming it's just a minor
change and won't have side-effects.

Test plan:
Make sure the tests pass on U20 and another OS

Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>