Bug 25651: Fix modification of authorised value
When modifying an existing authorised value, the category is not
correctly passed to the template and so it is not passed to the
POST request, which results in the authorised value switching to
category '', which effectively make it disappear from the interface.
This patch fixes that.

Signed-off-by: Sally <sally.healey@cheshiresharedservices.gov.uk>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit 4d003ad9cd5a0967f3235d687340a8b75d36acd3)

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>