Bug 25593: Terminology: Fix "There is no order for this biblio." on catalog detail page
This patch changes the phrase "There is no order for this biblio" to
"There is no order for this bibliographic record." on the catalog
detail page.

To test, apply the patch and view the bibliographic detail page for a
record which has no order associated with it. The text in the
"Acquisitions details" tab should be correct.

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>