Bug 25687: Switch Y/N in EDI accounts table for Yes and No for better translatability
This patch changes the list of EDI accounts shown in EDI administration
so that columns which previously used "Y" or "N" now show "Yes" or "No."

To test, apply the patch and go to Administration -> EDI accounts.

If necessary, add an EDI account. In the table of information about the
EDI account, the columns with information about quotes, orders,
invoices, responses, and auto-ordering should show "Yes" or "No" based
on the account's settings.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit 40aaf25409921b1414396bb1aa2146b63d432339)

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>